+++ .eslintrc.js
@@ -7,18 +7,22 @@ module.exports = {
     extends: [
         "plugin:vue/recommended",
         "eslint:recommended",
-        "prettier/vue",
+        'plugin:vue/vue3-recommended',
         "plugin:prettier/recommended",
     ],
     rules: {
         "vue/component-name-in-template-casing": ["error", "PascalCase"],
         "no-console": process.env.NODE_ENV === "production" ? "error" : "off",
         "no-debugger": process.env.NODE_ENV === "production" ? "error" : "off",
+        "vue/multi-word-component-names": "off"
+
     },
     globals: {
         $nuxt: true,
     },
     parserOptions: {
-        parser: "babel-eslint",
+        parser: "@babel/eslint-parser",
+        requireConfigFile: false,
     },
+
 };
