+++ components/MainLogo.vue
