+++ components/ResourcesSection.vue
@@ -31,12 +31,12 @@
 </template>
 
 <script>
-import ExternalInfoCard from "./ExternalInfoCard.vue";
+// import ExternalInfoCard from "./ExternalInfoCard.vue";
 export default {
-    components: { ExternalInfoCard },
-//     setup () {
-//     const store = useStore()
-//   },
+    // components: { ExternalInfoCard },
+    //     setup () {
+    //     const store = useStore()
+    //   },
     computed: {
         // resources() {
         //     return this.$store.getters["resources/getAllResources"];
